torchserve/requirements/requirements.cpu.txt
# facenet-pytorch===2.5.2
mmcv===1.6.0
torchserve===0.6.0
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.1
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
torch>=1.7.0,!=1.12.0
torchvision>=0.8.1,!=0.13.0
tqdm>=4.41.0
protobuf<4.21.3

# Logging -------------------------------------
tensorboard>=2.4.1
# wandb

# Plotting ------------------------------------
pandas>=1.1.4
seaborn>=0.11.0
# misc
ilock===1.0.3

# triangle_vod

timm===0.6.7
pytorch-lightning===1.7.2

# misc
ilock===1.0.3
einops===0.4.1

kornia===0.6.6
natsort===8.1.0